refactor: prepare modal ui set for implementation
